- Minor updates to the visual studio projects - Updates ProximityStrategyFactory s.t. default is to use the build in collision checking strategy
git-svn-id: https://svnsrv.sdu.dk/svn/RobWork/trunk@4056 494c09c7-a67d-49d4-9938-151fb1965be8

+++ RobWork/src/rw/kinematics/StateStructure.hpp
@@ -74,7 +74,7 @@ namespace rw { namespace kinematics {
          * lower than this number (and greater than or equal to zero).
          *
          */
-        int getMaxID() const { return _allDatas.size(); }
+        int getMaxID() const { return (int)_allDatas.size(); }
 
         /**
          * @brief adds a statedata to the frame tree and allocates memory
